create folders Exceptions and Responses and move the responses and exceptions files into them
