+++ TransbankPosSDK/Exceptions/CommonExceptions/TransbankException.cs
