+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankCloseException.cs
