+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankGetTotalsException.cs
