+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankLastSaleException.cs
