+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankMultiCodeDetailException.cs
