+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankMultiCodeLastSaleException.cs
