+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankMultiCodeSaleException.cs
