+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankRefundException.cs
