+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankSaleException.cs
