+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankSalesDetailException.cs
