+++ TransbankPosSDK/Exceptions/IntegradoExceptions/TransbankTotalsException.cs
