+++ TransbankPosSDK/Responses/IntegradoResponses/MultiCodeDetailResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Transbank.POS.IntegradoResponses
+namespace Transbank.POS.Responses.IntegradoResponses
 {
     public class MultiCodeDetailResponse : SaleResponse
     {
